+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/AuthnResponseResult.kt
@@ -12,7 +12,7 @@ data class AuthnResponseResult(
     val idTokenValidationResult: KmmResult<IdToken>?,
     val vpTokenValidationResult: KmmResult<VpTokenValidationResult>?,
     val request: AuthenticationRequestParameters?,
-) {
+) : DcApiResponseResult() {
     val state
         get() = request?.state
 }