+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/iso/Iso180137AnnexCRequestOptions.kt
@@ -3,6 +3,7 @@ package at.asitplus.wallet.lib.iso
 import at.asitplus.iso.DeviceRequest
 import at.asitplus.wallet.lib.RequestOptions
 
+@Deprecated("Use OpenId4VpRequestOptions and DcApiVerifier instead")
 data class Iso180137AnnexCRequestOptions(
     /**
      * Device request can be built using [CredentialPresentationRequestBuilder]