+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/iso/Iso180137AnnexCVerifiedPresentationResult.kt
@@ -2,9 +2,7 @@ package at.asitplus.wallet.lib.iso
 
 import at.asitplus.wallet.lib.data.IsoDocumentParsed
 
-/**
- * Validation results of ISO 18013-7 presentation
- */
+@Deprecated("Use Iso180137AnnexCWrapper and DcApiVerifier instead")
 data class Iso180137AnnexCVerifiedPresentationResult(
     val documents: Collection<IsoDocumentParsed>,
 )